requirements.txt
watchdog>=3.0.0
pillow>=10.0.0
pyautogui>=0.9.54
psutil>=5.9.0